Renamed services for clarity
Add types for condition summaries
